+++ cypress/integration/drag-svg.ts
@@ -1,10 +1,19 @@
+/**
+ * In this test suite there's two sets of each test, one without and one with the `layout` prop.
+ * This is because when a component has the layout prop applied, we apply drag to the bounding box
+ * and when it isn't, we apply it to the x/y transform.
+ *
+ * Descrepencies between the expected values in the two sets of tests are *something* to do with how
+ * pointer events are being resolved with Cypress, but a manual check will verify that both drag modes
+ * are working visually the same.
+ */
 describe("Drag SVG", () => {
     it("Drags the element by the defined distance", () => {
         cy.visit("?test=drag-svg")
             .get("[data-testid='draggable']")
             .wait(200)
             .trigger("pointerdown", 0, 0, { force: true })
-            .trigger("pointermove", 60, 60, { force: true }) // Gesture will start from first move past threshold
+            .trigger("pointermove", 10, 10, { force: true }) // Gesture will start from first move past threshold
             .wait(50)
             .trigger("pointermove", 200, 300, { force: true })
             .wait(50)
@@ -13,8 +22,8 @@ describe("Drag SVG", () => {
                 const draggable = $draggable[0] as HTMLDivElement
                 const { left, top } = draggable.getBoundingClientRect()
 
-                expect(left).to.equal(170)
-                expect(top).to.equal(270)
+                expect(left).to.equal(200)
+                expect(top).to.equal(300)
             })
     })
 
@@ -32,7 +41,7 @@ describe("Drag SVG", () => {
                 const draggable = $draggable[0] as HTMLDivElement
                 const { left, top } = draggable.getBoundingClientRect()
 
-                expect(left).to.equal(170)
+                expect(left).to.equal(190)
                 expect(top).to.equal(30)
             })
     })
@@ -52,7 +61,7 @@ describe("Drag SVG", () => {
                 const { left, top } = draggable.getBoundingClientRect()
 
                 expect(left).to.equal(30)
-                expect(top).to.equal(270)
+                expect(top).to.equal(290)
             })
     })
 
@@ -122,6 +131,148 @@ describe("Drag SVG", () => {
 
     it("Constraints as object: top left", () => {
         cy.visit("?test=drag-svg&left=-10&top=-10")
+            .wait(200)
+            .get("[data-testid='draggable']")
+            .trigger("pointerdown", 50, 50, { force: true })
+            .trigger("pointermove", 60, 60, { force: true }) // Gesture will start from first move past threshold
+            .wait(50)
+            .trigger("pointermove", 10, 10, { force: true })
+            .wait(50)
+            .trigger("pointerup", { force: true })
+            .should(($draggable: any) => {
+                const draggable = $draggable[0] as HTMLDivElement
+                const { left, top } = draggable.getBoundingClientRect()
+
+                expect(left).to.equal(20)
+                expect(top).to.equal(20)
+            })
+    })
+})
+
+describe("Drag SVG & Layout", () => {
+    it("Drags the element by the defined distance", () => {
+        cy.visit("?test=drag-svg&layout=true")
+            .get("[data-testid='draggable']")
+            .wait(200)
+            .trigger("pointerdown", 0, 0, { force: true })
+            .trigger("pointermove", 60, 60, { force: true }) // Gesture will start from first move past threshold
+            .wait(50)
+            .trigger("pointermove", 200, 300, { force: true })
+            .wait(50)
+            .trigger("pointerup", { force: true })
+            .should(($draggable: any) => {
+                const draggable = $draggable[0] as HTMLDivElement
+                const { left, top } = draggable.getBoundingClientRect()
+
+                expect(left).to.equal(170)
+                expect(top).to.equal(270)
+            })
+    })
+
+    it("Locks drag to x", () => {
+        cy.visit("?test=drag-svg&axis=x&layout=true")
+            .get("[data-testid='draggable']")
+            .wait(200)
+            .trigger("pointerdown", 50, 50, { force: true })
+            .trigger("pointermove", 60, 60, { force: true }) // Gesture will start from first move past threshold
+            .wait(50)
+            .trigger("pointermove", 200, 300, { force: true })
+            .wait(50)
+            .trigger("pointerup", { force: true })
+            .should(($draggable: any) => {
+                const draggable = $draggable[0] as HTMLDivElement
+                const { left, top } = draggable.getBoundingClientRect()
+
+                expect(left).to.equal(170)
+                expect(top).to.equal(30)
+            })
+    })
+
+    it("Locks drag to y", () => {
+        cy.visit("?test=drag-svg&axis=y&layout=true")
+            .get("[data-testid='draggable']")
+            .wait(200)
+            .trigger("pointerdown", 50, 50, { force: true })
+            .trigger("pointermove", 60, 60, { force: true }) // Gesture will start from first move past threshold
+            .wait(50)
+            .trigger("pointermove", 200, 300, { force: true })
+            .wait(50)
+            .trigger("pointerup", { force: true })
+            .should(($draggable: any) => {
+                const draggable = $draggable[0] as HTMLDivElement
+                const { left, top } = draggable.getBoundingClientRect()
+
+                expect(left).to.equal(30)
+                expect(top).to.equal(270)
+            })
+    })
+
+    // TODO: Figure out direction lock tests in Cypress
+    // it("Direction locks to x", () => {
+    //     cy.visit("?test=drag-svg&lock=true")
+    //         .wait(200)
+    //         .get("[data-testid='draggable']")
+    //         .trigger("pointerdown", 50, 50, { force: true })
+
+    //     cy.window()
+    //         .trigger("pointermove", 60, 50, { force: true }) // Gesture will start from first move past threshold
+    //         .wait(100)
+    //         .trigger("pointermove", 200, 0, { force: true })
+    //         .wait(50)
+    //         // .trigger("pointermove", 200, 200, { force: true })
+    //         // .wait(50)
+    //         .trigger("pointerup", { force: true })
+    //     cy.get("[data-testid='draggable']").should(($draggable: any) => {
+    //         const draggable = $draggable[0] as HTMLDivElement
+    //         const { left, top } = draggable.getBoundingClientRect()
+
+    //         expect(left).to.equal(170)
+    //         expect(top).to.equal(30)
+    //     })
+    // })
+
+    // it("Direction locks to y", () => {
+    //     cy.visit("?test=drag-svg&lock=true")
+    //         .get("[data-testid='draggable']")
+    //         .wait(200)
+    //         .trigger("pointerdown", 50, 50, { force: true })
+    //         .trigger("pointermove", 50, 60, { force: true }) // Gesture will start from first move past threshold
+    //         .wait(200)
+    //         .trigger("pointermove", 50, 200, { force: true })
+    //         .wait(50)
+    //         .trigger("pointermove", 200, 50, { force: true })
+    //         .wait(50)
+    //         .trigger("pointerup", { force: true })
+    //         .should(($draggable: any) => {
+    //             const draggable = $draggable[0] as HTMLDivElement
+    //             const { left, top } = draggable.getBoundingClientRect()
+
+    //             expect(left).to.equal(30)
+    //             expect(top).to.equal(270)
+    //         })
+    // })
+
+    it("Constraints as object: bottom right", () => {
+        cy.visit("?test=drag-svg&right=100&bottom=100&layout=true")
+            .wait(200)
+            .get("[data-testid='draggable']")
+            .trigger("pointerdown", 50, 50, { force: true })
+            .trigger("pointermove", 60, 60, { force: true }) // Gesture will start from first move past threshold
+            .wait(50)
+            .trigger("pointermove", 200, 200, { force: true })
+            .wait(50)
+            .trigger("pointerup", { force: true })
+            .should(($draggable: any) => {
+                const draggable = $draggable[0] as HTMLDivElement
+                const { left, top } = draggable.getBoundingClientRect()
+
+                expect(left).to.equal(90)
+                expect(top).to.equal(90)
+            })
+    })
+
+    it("Constraints as object: top left", () => {
+        cy.visit("?test=drag-svg&left=-10&top=-10&layout=true")
             .wait(200)
             .get("[data-testid='draggable']")
             .trigger("pointerdown", 50, 50, { force: true })